earlystopping: Refactor
