+++ pkg/earlystopping/types.go
