+++ test/scripts/build.sh
@@ -41,7 +41,7 @@ go build -o bin/dlkmanager github.com/kubeflow/katib/dlk/dlkmanager
 go build -o bin/katib-suggestion-grid github.com/kubeflow/katib/suggestion/grid
 go build -o bin/katib-suggestion-hyperband github.com/kubeflow/katib/suggestion/hyperband
 go build -o bin/katib-suggestion-random github.com/kubeflow/katib/suggestion/random
-go build -o bin/katib-earlystopping-medianstopping github.com/kubeflow/katib/earlystopping/medianstopping
+go build -o bin/katib-earlystopping-medianstopping github.com/kubeflow/katib/cmd/earlystopping/medianstopping
 go build -o bin/katib github.com/kubeflow/katib/cmd/cli
 #echo "building container in gcloud"
 #gcloud version
